refactor(chess): 메서드명 수정 및 매직넘버 상수화
