+++ src/main/java/chess/Application.java
@@ -1,5 +1,7 @@
 package chess;
 
+import chess.game.ChessGame;
+
 public class Application {
 
     public static void main(String[] args) {
