+++ src/main/java/chess/view/input/InputView.java
@@ -7,6 +7,8 @@ import java.util.stream.Collectors;
 
 public class InputView {
 
+    private static final String COMMAND_SPLIT_DELIMITER = " ";
+
     private static final Scanner scanner = new Scanner(System.in);
 
     private InputView() {
@@ -14,7 +16,7 @@ public class InputView {
     }
 
     public static List<String> inputCommand() {
-        return Arrays.stream(scanner.nextLine().split(" "))
+        return Arrays.stream(scanner.nextLine().split(COMMAND_SPLIT_DELIMITER))
                 .collect(Collectors.toList());
     }
 }
