+++ src/main/java/chess/view/output/OutputView.java
@@ -13,6 +13,7 @@ import java.util.stream.Collectors;
 public class OutputView {
 
     private static final String EMPTY_SYMBOL = ".";
+    private static final String SCORE_MESSAGE_FORMAT = "whiteScore : %f | blackScore : %f%n";
 
     private OutputView() {
         throw new AssertionError();
@@ -54,6 +55,6 @@ public class OutputView {
     }
 
     public static void printScore(double whiteTeamScore, double blackTeamScore) {
-        System.out.printf("whiteScore : %f | blackScore : %f%n", whiteTeamScore, blackTeamScore);
+        System.out.printf(SCORE_MESSAGE_FORMAT, whiteTeamScore, blackTeamScore);
     }
 }
